feat: add a redis_service option
With this option, you can have a dedicated redis instance for your
plugin. It can be used for cache for example but not for storing states
(it's the job of the mfbase module). So you have hard limits on memory
usage for this instance, unix-socket only usage, no persistance. A
special env var REDIS_SOCKET_UNIX_SOCKET_PATH is injected into your
plugin environnement with the fullpath of your plugin redis unix socket.

Close #5